backend/requirements.txt
pyarrow==9.0.0
fastapi~=0.75.2
pandas~=1.3.5
requests~=2.27.1
python-multipart==0.0.5
uvicorn~=0.17.6
lightgbm==3.2.1
pyyaml==6.0
pyyaml~=6.0
scikit-learn~=1.0.2
optuna~=3.0.2
joblib~=1.1.0
numpy~=1.20.3